put class Node in a seperate file -> Node.cs
