+++ PikachuClassic/Grid/GridManager.cs
@@ -353,26 +353,6 @@ namespace PikachuClassic
         // Hết Bảo
         */
 
-        // Bảo, class Node phục vụ cho việc đồ thị hóa grid và lập danh sách kề
-        public class Node
-        {
-            public PictureBox PictureBox { get; }
-            public List<Node> Neighbors { get; }
-            public bool IsBoundary { get; } // Thuộc tính để xác định node viền ngoài
-
-            public Node(PictureBox pictureBox, bool isBoundary = false)
-            {
-                PictureBox = pictureBox;
-                Neighbors = new List<Node>();
-                IsBoundary = isBoundary; // Khởi tạo thuộc tính IsBoundary
-            }
-
-            public void AddNeighbor(Node neighbor)
-            {
-                Neighbors.Add(neighbor);
-            }
-        }
-
         // Bảo, dùng để loại node đã bị vô hiệu hóa khi chọn 2 hình giống nhau + cập nhật lại danh sách kề
         private void RemoveNodes(PictureBox firstBox, PictureBox secondBox)
         {
